requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
pandas>=1.3.0

# GUI Framework
PyQt5>=5.15.0
PyOpenGL>=3.1.0

# Hardware interfaces
pyserial>=3.5
cortex-python>=2.0.0  # For Emotiv EEG (fallback)
websockets>=10.0      # For WebSocket Cortex client
aiohttp>=3.8.0        # For async HTTP operations

# Machine Learning
tensorflow>=2.8.0
scikit-learn>=1.0.0
torch>=1.10.0
torchvision>=0.11.0

# Data handling
h5py>=3.6.0
PyYAML>=6.0

# Utilities
tqdm>=4.62.0
colorama>=0.4.4
click>=8.0.0

# Development
pytest>=6.2.0
black>=22.0.0
flake8>=4.0.0